+++ accounts/admin.py
@@ -1,3 +1,5 @@
 from django.contrib import admin
+from accounts.models import Car
 
 # Register your models here.
+admin.site.register(Car)
