+++ sales/admin.py
@@ -1,8 +1,7 @@
 from django.contrib import admin
-from sales.models import Address, Acquisition, Item, Car
+from sales.models import Address, Acquisition, Item
 
 # Register your models here.
 admin.site.register(Address)
 admin.site.register(Acquisition)
 admin.site.register(Item)
-admin.site.register(Car)
